+++ saleor/graphql/core/mutations.py
@@ -639,6 +639,7 @@ class ModelDeleteMutation(ModelMutation):
         # After the instance is deleted, set its ID to the original database's
         # ID so that the success response contains ID of the deleted object.
         instance.id = db_id
+        cls.post_save_action(info, instance, None)
         return cls.success_response(instance)
