+++ saleor/graphql/webhook/enums.py
@@ -51,6 +51,12 @@ WEBHOOK_EVENT_DESCRIPTION = {
     WebhookEventAsyncType.PRODUCT_VARIANT_CREATED: "A new product variant is created.",
     WebhookEventAsyncType.PRODUCT_VARIANT_UPDATED: "A product variant is updated.",
     WebhookEventAsyncType.PRODUCT_VARIANT_DELETED: "A product variant is deleted.",
+    WebhookEventAsyncType.SHIPPING_PRICE_CREATED: "A new shipping price is created.",
+    WebhookEventAsyncType.SHIPPING_PRICE_UPDATED: "A shipping price is updated.",
+    WebhookEventAsyncType.SHIPPING_PRICE_DELETED: "A shipping price is deleted.",
+    WebhookEventAsyncType.SHIPPING_ZONE_CREATED: "A new shipping zone is created.",
+    WebhookEventAsyncType.SHIPPING_ZONE_UPDATED: "A shipping zone is updated.",
+    WebhookEventAsyncType.SHIPPING_ZONE_DELETED: "A shipping zone is deleted.",
     WebhookEventAsyncType.INVOICE_REQUESTED: "An invoice for order requested.",
     WebhookEventAsyncType.INVOICE_DELETED: "An invoice is deleted.",
     WebhookEventAsyncType.INVOICE_SENT: "Invoice has been sent.",
