+++ saleor/plugins/manager.py
@@ -62,6 +62,7 @@ if TYPE_CHECKING:
         ProductVariant,
     )
     from ..shipping.interface import ShippingMethodData
+    from ..shipping.models import ShippingMethod, ShippingZone
     from ..translation.models import Translation
     from ..warehouse.models import Stock
     from .base_plugin import BasePlugin
@@ -797,6 +798,42 @@ class PluginsManager(PaymentInterface):
     def category_deleted(self, category: "Category"):
         default_value = None
         return self.__run_method_on_plugins("category_deleted", default_value, category)
+
+    def shipping_price_created(self, shipping_method: "ShippingMethod"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_price_created", default_value, shipping_method
+        )
+
+    def shipping_price_updated(self, shipping_method: "ShippingMethod"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_price_updated", default_value, shipping_method
+        )
+
+    def shipping_price_deleted(self, shipping_method: "ShippingMethod"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_price_deleted", default_value, shipping_method
+        )
+
+    def shipping_zone_created(self, shipping_zone: "ShippingZone"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_zone_created", default_value, shipping_zone
+        )
+
+    def shipping_zone_updated(self, shipping_zone: "ShippingZone"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_zone_updated", default_value, shipping_zone
+        )
+
+    def shipping_zone_deleted(self, shipping_zone: "ShippingZone"):
+        default_value = None
+        return self.__run_method_on_plugins(
+            "shipping_zone_deleted", default_value, shipping_zone
+        )
 
     def initialize_payment(
         self, gateway, payment_data: dict, channel_slug: str
