+++ saleor/webhook/event_types.py
@@ -67,6 +67,14 @@ class WebhookEventAsyncType:
     PAGE_UPDATED = "page_updated"
     PAGE_DELETED = "page_deleted"
 
+    SHIPPING_PRICE_CREATED = "shipping_price_created"
+    SHIPPING_PRICE_UPDATED = "shipping_price_updated"
+    SHIPPING_PRICE_DELETED = "shipping_price_deleted"
+
+    SHIPPING_ZONE_CREATED = "shipping_zone_created"
+    SHIPPING_ZONE_UPDATED = "shipping_zone_updated"
+    SHIPPING_ZONE_DELETED = "shipping_zone_deleted"
+
     TRANSLATION_CREATED = "translation_created"
     TRANSLATION_UPDATED = "translation_updated"
 
@@ -111,6 +119,12 @@ class WebhookEventAsyncType:
         PAGE_CREATED: "Page Created",
         PAGE_UPDATED: "Page Updated",
         PAGE_DELETED: "Page Deleted",
+        SHIPPING_PRICE_CREATED: "Shipping price created",
+        SHIPPING_PRICE_UPDATED: "Shipping price updated",
+        SHIPPING_PRICE_DELETED: "Shipping price deleted",
+        SHIPPING_ZONE_CREATED: "Shipping zone created",
+        SHIPPING_ZONE_UPDATED: "Shipping zone updated",
+        SHIPPING_ZONE_DELETED: "Shipping zone deleted",
         TRANSLATION_CREATED: "Create translation",
         TRANSLATION_UPDATED: "Update translation",
     }
@@ -156,6 +170,12 @@ class WebhookEventAsyncType:
         (PAGE_CREATED, DISPLAY_LABELS[PAGE_CREATED]),
         (PAGE_UPDATED, DISPLAY_LABELS[PAGE_UPDATED]),
         (PAGE_DELETED, DISPLAY_LABELS[PAGE_DELETED]),
+        (SHIPPING_PRICE_CREATED, DISPLAY_LABELS[SHIPPING_PRICE_CREATED]),
+        (SHIPPING_PRICE_UPDATED, DISPLAY_LABELS[SHIPPING_PRICE_UPDATED]),
+        (SHIPPING_PRICE_DELETED, DISPLAY_LABELS[SHIPPING_PRICE_DELETED]),
+        (SHIPPING_ZONE_CREATED, DISPLAY_LABELS[SHIPPING_ZONE_CREATED]),
+        (SHIPPING_ZONE_UPDATED, DISPLAY_LABELS[SHIPPING_ZONE_UPDATED]),
+        (SHIPPING_ZONE_DELETED, DISPLAY_LABELS[SHIPPING_ZONE_DELETED]),
         (TRANSLATION_CREATED, DISPLAY_LABELS[TRANSLATION_CREATED]),
         (TRANSLATION_UPDATED, DISPLAY_LABELS[TRANSLATION_UPDATED]),
     ]
@@ -202,6 +222,12 @@ class WebhookEventAsyncType:
         PAGE_CREATED: PagePermissions.MANAGE_PAGES,
         PAGE_UPDATED: PagePermissions.MANAGE_PAGES,
         PAGE_DELETED: PagePermissions.MANAGE_PAGES,
+        SHIPPING_PRICE_CREATED: ShippingPermissions.MANAGE_SHIPPING,
+        SHIPPING_PRICE_UPDATED: ShippingPermissions.MANAGE_SHIPPING,
+        SHIPPING_PRICE_DELETED: ShippingPermissions.MANAGE_SHIPPING,
+        SHIPPING_ZONE_CREATED: ShippingPermissions.MANAGE_SHIPPING,
+        SHIPPING_ZONE_UPDATED: ShippingPermissions.MANAGE_SHIPPING,
+        SHIPPING_ZONE_DELETED: ShippingPermissions.MANAGE_SHIPPING,
         TRANSLATION_CREATED: SitePermissions.MANAGE_TRANSLATIONS,
         TRANSLATION_UPDATED: SitePermissions.MANAGE_TRANSLATIONS,
     }
@@ -317,6 +343,12 @@ SUBSCRIBABLE_EVENTS = [
     WebhookEventAsyncType.PAGE_CREATED,
     WebhookEventAsyncType.PAGE_UPDATED,
     WebhookEventAsyncType.PAGE_DELETED,
+    WebhookEventAsyncType.SHIPPING_PRICE_CREATED,
+    WebhookEventAsyncType.SHIPPING_PRICE_UPDATED,
+    WebhookEventAsyncType.SHIPPING_PRICE_DELETED,
+    WebhookEventAsyncType.SHIPPING_ZONE_CREATED,
+    WebhookEventAsyncType.SHIPPING_ZONE_UPDATED,
+    WebhookEventAsyncType.SHIPPING_ZONE_DELETED,
     WebhookEventAsyncType.TRANSLATION_CREATED,
     WebhookEventAsyncType.TRANSLATION_UPDATED,
 ]
